requirements.txt
pypdf>=4.0
python-docx>=1.1
sentence-transformers>=3.0
chromadb>=0.5
ollama>=0.3
streamlit>=1.35
